app done

+++ deleteList.py
@@ -24,34 +24,34 @@ class DeleteListHandler(webapp2.RequestHandler):
 			access_link = users.create_logout_url('/')
 			
 			# Get receipt id
-			try:
-				idShoppingList = self.request.get('idShoppingList')
-			except:
+			idShoppingList = self.request.get('idShoppingList', '')
+
+			if idShoppingList == '':
 				self.redirect('/error?error=La lista de la compra no existe :(')
 				return
-
-			# Get query from database
-			try:
+			else:
+				# Get query from database
 				shoppingList = ndb.Key(urlsafe = idShoppingList).get()
-			except:
-				self.redirect('/error?error=La lista de la compra no existe :(')
-				return
-			
-			# Delete id
-			shoppingList.key.delete()
 
-			time.sleep(1)
+				if shoppingList == None:
+					self.redirect('/error?error=La lista de la compra no existe :(')
+					return
+				else:
+					# Delete id
+					shoppingList.key.delete()
 
-			# Get query from database
-			shoppingLists = ShoppingList.query(ShoppingList.user == user.user_id()).order(-ShoppingList.date)
+					time.sleep(1)
 
-			template_values = {
-				'user_name' : user_name,
-				'access_link' : access_link,
-				'shoppingLists' : shoppingLists
-			}
+					# Get query from database
+					shoppingLists = ShoppingList.query(ShoppingList.user == user.user_id()).order(-ShoppingList.date)
 
-			template = JINJA_ENVIRONMENT.get_template('showList.html')
-			self.response.write(template.render(template_values));
+					template_values = {
+						'user_name' : user_name,
+						'access_link' : access_link,
+						'shoppingLists' : shoppingLists
+					}
+
+					template = JINJA_ENVIRONMENT.get_template('showList.html')
+					self.response.write(template.render(template_values));
 		else:
 			self.redirect('/')
